refactor: ProfilePageを200行以内にリファクタリング
コーディング規約準拠のため、タブコンテンツを分離:
- ProfilePage.tsx: 304行 → 88行
- PostsTab.tsx: 128行（新規作成）
- GrowthRecordsTab.tsx: 127行（新規作成）

単一責任原則を遵守し、各コンポーネントが独立した責務を持つように改善。
